🧹 rename gcp provider to google provider
