+++ motor/providers/google/gcloud_config.go
@@ -1,4 +1,4 @@
-package gcp
+package google
 
 import (
 	"encoding/json"
