+++ motor/providers/google/gcloud_config_test.go
@@ -1,4 +1,4 @@
-package gcp_test
+package google_test
 
 import (
 	"bytes"
@@ -7,14 +7,14 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.mondoo.com/cnquery/motor/providers/gcp"
+	"go.mondoo.com/cnquery/motor/providers/google"
 )
 
 func TestParseGcloudConfig(t *testing.T) {
 	data, err := os.ReadFile("./testdata/gcloud_config.json")
 	require.NoError(t, err)
 
-	config, err := gcp.ParseGcloudConfig(bytes.NewReader(data))
+	config, err := google.ParseGcloudConfig(bytes.NewReader(data))
 	require.NoError(t, err)
 
 	assert.Equal(t, "mondoo-abc-12345", config.Configuration.Properties.Core.Project)
