+++ motor/providers/google/organization.go
@@ -1,4 +1,4 @@
-package gcp
+package google
 
 import (
 	"context"
