+++ motor/providers/google/testdata/gcloud_config.json
